requirements.txt
pyyaml
flask
flask-socketio